indexical::bitset

Module rustc

The Rust compiler’s BitSet.

Re-exports

pub extern crate rustc_index;

Structs

ChunkedBitSet

A fixed-size bitset type with a partially dense, partially sparse representation. The bitset is broken into chunks, and chunks that are all zeros or all ones are represented and handled very efficiently.

DenseBitSet

A fixed-size bitset type with a dense representation.

Enums

MixedBitIter

MixedBitSet

A bitset with a mixed representation, using DenseBitSet for small and medium bitsets, and ChunkedBitSet for large bitsets, i.e. those with enough bits for at least two chunks. This is a good choice for many bitsets that can have large domain sizes (e.g. 5000+).

Type Aliases

ArcIndexMatrix

IndexMatrix specialized to the bit_set::BitSet implementation with the ArcFamily.

ArcIndexSet

IndexSet specialized to the bit_set::BitSet implementation with the ArcFamily.

IndexMatrix

IndexMatrix specialized to the bit_set::BitSet implementation.

IndexSet

IndexSet specialized to the bit_set::BitSet implementation.

RefIndexMatrix

IndexMatrix specialized to the bit_set::BitSet implementation with the RefFamily.

RefIndexSet

IndexSet specialized to the bit_set::BitSet implementation with the RefFamily.

RustcBitSet

A bitset specialized to usize indices.