Struct indexical::impls::bit_set::BitSet

pub struct BitSet<T> { /* private fields */ }

A fixed-size bitset type with a dense representation.

NOTE: Use GrowableBitSet if you need support for resizing after creation.

T is an index type, typically a newtyped usize wrapper, but it can also just be usize.

All operations that involve an element will panic if the element is equal to or greater than the domain size. All operations that involve two bitsets will panic if the bitsets have differing domain sizes.