indexical

Module map

Map-like collections for indexed keys.

Structs

DenseIndexMap

A mapping from indexed keys to values, implemented densely with a vector.

SparseIndexMap

A mapping from indexed keys to values, implemented sparsely with a hash map.

Type Aliases

DenseArcIndexMap

DenseIndexMap specialized to the ArcFamily.

DenseRcIndexMap

DenseIndexMap specialized to the RcFamily.

DenseRefIndexMap

DenseIndexMap specialized to the RefFamily.

SparseArcIndexMap

SparseIndexMap specialized to the ArcFamily.

SparseRcIndexMap

SparseIndexMap specialized to the RcFamily.

SparseRefIndexMap

SparseIndexMap specialized to the RefFamily.