indexical::map

Struct DenseIndexMap

pub struct DenseIndexMap<'a, K: IndexedValue + 'a, V, P: PointerFamily<'a>>(/* private fields */);

A mapping from indexed keys to values, implemented densely with a vector.

This is more time-efficient than the SparseIndexMap for lookup, but it consumes more memory for missing elements.

Implementations

impl<'a, K, V, P> DenseIndexMap<'a, K, V, P>
where K: IndexedValue + 'a, P: PointerFamily<'a>,

pub fn new(domain: &P::Pointer<IndexedDomain<K>>) -> Self

Constructs a new map with an initial element of mk_elem(i) for each i in domain.

pub fn get<M>(&self, idx: impl ToIndex<K, M>) -> Option<&V>

Returns an immutable reference to a value for a given key if it exists.

pub fn get_mut<M>(&mut self, idx: impl ToIndex<K, M>) -> Option<&mut V>

Returns a mutable reference to a value for a given key if it exists.

pub unsafe fn get_unchecked<M>(&self, idx: impl ToIndex<K, M>) -> &V

Returns a reference to a value for a given key.

Safety

This function has undefined behavior if key is not in self.

pub unsafe fn get_unchecked_mut<M>(&mut self, idx: impl ToIndex<K, M>) -> &mut V

Returns a mutable reference to a value for a given key.

Safety

This function has undefined behavior if key is not in self.

pub fn insert<M>(&mut self, idx: impl ToIndex<K, M>, value: V)

Inserts the key/value pair into self.

pub fn values(&self) -> impl Iterator<Item = &V>

Returns an iterator over the values of the map.