indexical::map

Struct SparseIndexMap

pub struct SparseIndexMap<'a, K: IndexedValue + 'a, V, P: PointerFamily<'a>> { /* private fields */ }

A mapping from indexed keys to values, implemented sparsely with a hash map.

This is more memory-efficient than the DenseIndexMap with a small number of keys.

Implementations

impl<'a, K, V, P> SparseIndexMap<'a, K, V, P>
where K: IndexedValue + 'a, P: PointerFamily<'a>,

pub fn new(domain: &P::Pointer<IndexedDomain<K>>) -> Self

Constructs an empty map within the given domain.

pub fn get<M>(&self, key: impl ToIndex<K, M>) -> Option<&V>

Returns an immutable reference to a value for a given key if it exists.

pub fn get_mut<M>(&mut self, key: impl ToIndex<K, M>) -> Option<&mut V>

Returns a mutable reference to a value for a given key if it exists.

pub unsafe fn get_unchecked<M>(&self, key: impl ToIndex<K, M>) -> &V

Returns a reference to a value for a given key.

Safety

This function has undefined behavior if key is not in self.

pub unsafe fn get_unchecked_mut<M>(&mut self, key: impl ToIndex<K, M>) -> &mut V

Returns a mutable reference to a value for a given key.

Safety

This function has undefined behavior if key is not in self.

pub fn insert<M>(&mut self, key: impl ToIndex<K, M>, value: V)

Inserts the key/value pair into self.

pub fn values(&self) -> impl ExactSizeIterator<Item = &V>

Returns an iterator over the values of the map.

pub fn entry<M>(&mut self, key: impl ToIndex<K, M>) -> Entry<'_, K::Index, V>

Returns a mutable entry into the map for the given key.

pub fn len(&self) -> usize

Returns the number of entries in the map.

pub fn is_empty(&self) -> bool

Returns true if the map has no elements.