indexical

Struct IndexSet

pub struct IndexSet<'a, T: IndexedValue + 'a, S, P: PointerFamily<'a>> { /* private fields */ }

An unordered collections of Ts, implemented with a bit-set.

Implementations

impl<'a, T, S, P> IndexSet<'a, T, S, P>
where T: IndexedValue + 'a, S: BitSet, P: PointerFamily<'a>,

pub fn new(domain: &P::Pointer<IndexedDomain<T>>) -> Self

Creates an empty index set.

pub fn indices(&self) -> impl Iterator<Item = T::Index>

Returns an iterator over all the indices contained in self.

pub fn iter(&self) -> impl Iterator<Item = &T>

Returns an iterator over all the objects contained in self.

pub fn iter_enumerated(&self) -> impl Iterator<Item = (T::Index, &T)>

Returns an iterator over the pairs of indices and objects contained in self.

pub fn contains<M>(&self, index: impl ToIndex<T, M>) -> bool

Returns true if index is contained in self.

pub fn len(&self) -> usize

Returns the number of elements in self.

pub fn is_empty(&self) -> bool

Return true if self has no elements.

pub fn is_superset(&self, other: &IndexSet<'a, T, S, P>) -> bool

Returns true if every element in other is also in self.

pub fn insert<M>(&mut self, elt: impl ToIndex<T, M>) -> bool

Adds the element elt to self, returning true if self changed.

pub fn remove<M>(&mut self, elt: impl ToIndex<T, M>) -> bool

Removes the element elt from self, returning true if self changed.

pub fn union(&mut self, other: &IndexSet<'a, T, S, P>)

Adds each element of other to self.

pub fn union_changed(&mut self, other: &IndexSet<'a, T, S, P>) -> bool

Adds each element of other to self, returning true if self changed.

pub fn subtract(&mut self, other: &IndexSet<'a, T, S, P>)

Removes every element of other from self.

pub fn subtract_changed(&mut self, other: &IndexSet<'a, T, S, P>) -> bool

Removes every element of other from self, returning true if self changed.

pub fn intersect(&mut self, other: &IndexSet<'a, T, S, P>)

Removes every element of self not in other.

pub fn intersect_changed(&mut self, other: &IndexSet<'a, T, S, P>) -> bool

Removes every element of self not in other, returning true if self changed.

pub fn insert_all(&mut self)

Adds every element of the domain to self.

pub fn clear(&mut self)

Removes every element from self.

pub fn inner(&self) -> &S

Returns a reference to the inner set.