Trait indexical::PointerFamily

pub trait PointerFamily {
    type Pointer<T: 'static>: Deref<Target = T> + Clone;
}

Abstraction over smart pointers with 'static interiors.

Used so to make the indexical data structures generic with respect to choice of Rc or Arc (or your own clonable smart pointer!).

Required Associated Types

type Pointer<T: 'static>: Deref<Target = T> + Clone

Pointer type for a given family.

Object Safety

This trait is not object safe.