indexical::vec

Struct IndexVec

pub struct IndexVec<'a, K: IndexedValue + 'a, V, P: PointerFamily<'a>> { /* private fields */ }

A fixed-sized vector with one value for each key in the domain.

Implementations

impl<'a, K, V, P> IndexVec<'a, K, V, P>
where K: IndexedValue + 'a, P: PointerFamily<'a>, V: Clone,

pub fn from_elem(elem: V, domain: &P::Pointer<IndexedDomain<K>>) -> Self

Constructs a new vector where each index maps to the initial element elem.

impl<'a, K, V, P> IndexVec<'a, K, V, P>
where K: IndexedValue + 'a, P: PointerFamily<'a>,

pub fn from_fn( f: impl FnMut(K::Index) -> V, domain: &P::Pointer<IndexedDomain<K>>, ) -> Self

Constructs a new vector where each index maps to the output of f(index).

pub fn get<M>(&self, idx: impl ToIndex<K, M>) -> &V

Returns an immutable reference to a value for a given index.

pub fn get_mut<M>(&mut self, idx: impl ToIndex<K, M>) -> &mut V

Returns a mutable reference to a value for a given index.

pub fn iter( &self, ) -> impl DoubleEndedIterator<Item = &V> + ExactSizeIterator<Item = &V>

Returns an iterator over immutable references to the values.

pub fn iter_mut( &mut self, ) -> impl DoubleEndedIterator<Item = &mut V> + ExactSizeIterator<Item = &mut V>

Returns an iterator over mutable references to the values.

pub fn as_slice(&self) -> &[V]

Returns the underlying vector as a slice.

pub fn as_slice_mut(&mut self) -> &mut [V]

Returns the underlying vector as a mutable slice.

pub fn get_disjoint_mut<const N: usize>( &mut self, indices: [K::Index; N], ) -> Result<[&mut V; N], GetDisjointMutError>

Returns multiple mutable references to disjoint indices, or a GetDisjointMutError if not disjoint or in-bounds.