indexical::impls
Module bit_set

Structs

BitIter

BitMatrix

A fixed-size 2D bit matrix type with a dense representation.

BitSet

A fixed-size bitset type with a dense representation.

ChunkedBitIter

ChunkedBitSet

A fixed-size bitset type with a partially dense, partially sparse representation. The bitset is broken into chunks, and chunks that are all zeros or all ones are represented and handled very efficiently.

FiniteBitSet

A fixed-sized bitset type represented by an integer type. Indices outwith than the range representable by T are considered set.

GrowableBitSet

A resizable bitset type with a dense representation.

SparseBitMatrix

A fixed-column-size, variable-row-size 2D bit matrix with a moderately sparse representation.

Enums

MixedBitIter

MixedBitSet

A bitset with a mixed representation, using BitSet for small and medium bitsets, and ChunkedBitSet for large bitsets, i.e. those with enough bits for at least two chunks. This is a good choice for many bitsets that can have large domain sizes (e.g. 5000+).

Traits

BitRelations

FiniteBitSetTy

Integral type used to represent the bit set.