Trait indexical::impls::bit_set::FiniteBitSetTy

pub trait FiniteBitSetTy: BitAnd<Self, Output = Self> + BitAndAssign<Self> + BitOrAssign<Self> + Clone + Copy + Shl<Self> + Not<Output = Self> + PartialEq<Self> + Sized {
    const DOMAIN_SIZE: u32;
    const FILLED: Self;
    const EMPTY: Self;
    const ONE: Self;
    const ZERO: Self;

    // Required methods
    fn checked_shl(self, rhs: u32) -> Option<Self>;
    fn checked_shr(self, rhs: u32) -> Option<Self>;
}

Integral type used to represent the bit set.

Required Associated Constants

const DOMAIN_SIZE: u32

Size of the domain representable by this type, e.g. 64 for u64.

const FILLED: Self

Value which represents the FiniteBitSet having every bit set.

const EMPTY: Self

Value which represents the FiniteBitSet having no bits set.

const ONE: Self

Value for one as the integral type.

const ZERO: Self

Value for zero as the integral type.

Required Methods

fn checked_shl(self, rhs: u32) -> Option<Self>

Perform a checked left shift on the integral type.

fn checked_shr(self, rhs: u32) -> Option<Self>

Perform a checked right shift on the integral type.