indexical
Trait PointerFamily

pub trait PointerFamily {
    type Pointer<T: 'static>: Deref<Target = T> + Clone;
}

Abstraction over smart pointers with 'static interiors.

Used so to make the indexical data structures generic with respect to choice of Rc or Arc (or your own clonable smart pointer!).

Required Associated Types

type Pointer<T: 'static>: Deref<Target = T> + Clone

Pointer type for a given family.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.